% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{query_sparql}
\alias{query_sparql}
\title{Query a SPARQL endpoint}
\usage{
query_sparql(sparql_query, endpoint, format = c("simple", "smart"))
}
\arguments{
\item{sparql_query}{character, the SPARQL query}

\item{endpoint}{character, the SPARQL endpoint}

\item{format}{"simple" uses CSV and returns pure character data frame, while
"smart" fetches JSON-formatted data and returns a data frame with datetime
columns converted to \code{POSIXct}}
}
\value{
A tibble data frame
}
\description{
Makes an HTTP POST request to a SPARQL endpoint with the given SPARQL query.
}
\examples{
sparql_query <- "SELECT
  ?softwareVersion ?publicationDate
WHERE {
  BIND(wd:Q206904 AS ?R)
  ?R p:P348 [
    ps:P348 ?softwareVersion;
    pq:P577 ?publicationDate
  ] .
}"
query_sparql(sparql_query, endpoint = "https://query.wikidata.org/sparql", format="smart")
}
\references{
The code here was adapted from query.R in the WikidataQueryServiceR
package, developed originally by \href{https://github.com/bearloga}{Mikhail Popov}.
}
